Wait for ClickHouse to report ready before parsing ports
This changes the logic for parsing the server ports that ClickHouse
listens on, so that it waits until we are certain that they've been
completely written. This fixes a few flaky tests, currently including at
least #4779, #4972, and #5180. There have been others in the past, which
we've addressed with less-complete solutions that only narrow the race
window. This should eliminate it.