[reconfigurator] add decommissioned sleds to the planning input (#6644)
With zone cleanup, part of the logic depends on knowing whether
particular sleds are decommissioned. Previously we weren't returning
such sleds at all, which means that the planner is unable to distinguish
between the cases where the sled is decommissioned and the sled is not
present in the input at all. (It would be really bad for zone cleanup to
remove zones if something went horribly wrong!)

There is the potential for some fallout from decommissioned entries
being present in the sled map -- I've tried to guard it best I can by
changing the lookup API making it accept a filter. An alternative would
be to store the set of decommissioned sleds separately, but that seems
worse long-term.

The change identified a small issue with the reconfigurator CLI --
previously, we'd be okay overwriting decommissioned sleds. That case
produces a soft error now.