Rename static route local_pref to rib_priority
This is the omicron half of maghemite#359. This renames the `local_pref`
field of static routes to `rib_priority` to avoid overloading the name
of a standardized and well-known BGP attribute (Local Preference). This
change shrinks the size of the field from u32 -> u8 and unwraps the
Option<T> before shipping routes to mgd, aligning with the updated API.
This also brings omicron back into sync with maghemite.

Fixes: #6711

Signed-off-by: Trey Aspelund <trey@oxidecomputer.com>