Store NVMe firmware metadata in inventory collection (#6275)
With #5706 merged and later reenabled in #6026 we now want to store this
data in CRDB so that it can be used by blueprints and reconfigurator.

`nexus_types::inventory::PhysicalDisk` now has a `firmware` field that
is exposed as an enum called `PhysicalDiskFirmware`. Today we only have
`Unknown` and `Nvme` variants but this can later be expanded upon if we
ever start using additional disk types such as SCSI/SATA etc.

Finally I have included an updated version of `omdb db inventory
physical-disks` that will include the currently active firmware as well
as the next active firmware if a new version has been staged.