add `DatasetName::mountpoint` function; fix inconsistent mountpoints for nexus-managed datasets (#6670)
While working in related code I spotted a bug in the `ensure_dataset`
handler: the mountpoint for encrypted datasets is not under the `crypt`
directory. This wouldn't keep the dataset from being encrypted since it
was still had the `crypt` dataset as its parent, but it would make
mountpoints different depending on whether it was created directly by
Sled Agent or created by request of Nexus.

`DatasetName` now has a method for getting the mountpoint which handles
adding the `crypt` path segment if needed.