quasi-lock dependencies in omicron-common job (#6692)
"fixes" https://github.com/oxidecomputer/omicron/issues/6691 by stealing
the checked-in Cargo.lock when building the standalone `omicron-common`
crate created for this job. this lets us keep the specific refs of git
dependencies that are `ref="main"` and similar, which might otherwise
resolve to different commits than the rest of the tree (and its
`omicron-common`) can actually build with.

it'd be nice to have fewer git dependencies using branch names for refs,
either by using a cratesio crate or specific commit or _something_, but
this seems to make things a bit less brittle in the mean time.