sled agent: don't special-case vmm-not-present handling for requests to stop (#6698)
When sled agent receives a request to stop a VMM that's not in the
agent's VMM table, return `NoSuchVmm` instead of succeeding. This allows
users manually to recover an instance that was Running prior to a sled
reboot but hasn't yet been moved to Failed by the instance watcher.

Tested manually as follows:

1. Modify sled agent's VMM worker loop so that it doesn't publish VMM
state before exiting; this is needed so that manually unregistering an
instance from a sled doesn't cause it to go to Stopped
2. Launch a dev cluster with both (1) and the change in this PR.
3. Start an instance, then send an HTTP DELETE to sled agent's internal
API to forcibly unregister the VMM.
4. Observe that the instance remains Running in the console.
5. Stop the instance; observe that the "not found, going to Failed"
message is displayed and that the instance then goes to Failed.

Fixes #4511.