hyper v1: work around the current maghemite / omicron lag (#6706)
This is kind of gross.

Omicron currently is out of sync with maghemite, see #6693 and the two
maghemite pushes that require synchronization:
https://github.com/oxidecomputer/maghemite/pull/359 and
https://github.com/oxidecomputer/maghemite/pull/360. I'd like to make
forward progress with the hyper v1 migration, and that's blocked on
updating the omicron dependency of maghemite which pulls in old hyper
(reqwest, progenitor, etc). This gets pull into other repos via
omicron-common.

It's worth noting that this circular arrangement seems lousy. The
"vassal crates" (crucible, propolis, maghemite, (and dendrite to a
lesser degree)) depend on omicron-common, but omicron-common pulls in
maghemite's `mg-admin-client` which in turn pulls in... lots...
including progenitor, hyper, reqwest, etc.

My goal here is to have a temporary dependency on a branch of maghemite
(https://github.com/oxidecomputer/maghemite/pull/378); once #6693 is
integrated, we can then pin the dependency on maghemite's HEAD rev.