Prune dep reqs for external nexus-client consumers
Until now, a dependency upon nexus-client meant dragging in
omicron-common as well, since consumers internal to omicron benefit from
the convenience of those omicron-native types being substituted in as
part of the progenitor client generation.  For external nexus-client
consumers, this is a hindrance rather than a convenience - a vast number
of transitive dependencies get dragged in with omicron-common, none of
which are of value to such a consumer.

This introduces the "omicron-internal" flag for nexus-client, as well as
a few other crates.  When enabled, it preserves the legacy
type-replacement behavior, maintaining the status quo for consumers
inside of omicron.  External consumers can elide the feature, and will
instead rely on the progenitor-generated types.

The "omicron-internal" concept extends to the internal-dns crate as
well, where the Resolver functionality is necessary for external
consumers to locate the nexus service.  Internal consumers make use of
the richer network-related primitives for tasks such as zone
configuration, and thus require the omicron-common dependency through
assertion of the "omicron-internal" feature flag.