Remove no-longer-needed internal NTP configuration options (#6657)
This grew a bit bigger than I expected. The main goal is to fix #6261:
now that R10 is out the door, we no longer need to tell each internal
NTP zone the specific names of all boundary NTP zones. That grew into a
few other cleanup operations:

* I removed the `ntp_servers`, `dns_servers`, and `domain` fields from
`OmicronZoneType::InternalNtp`. The only thing it needs now is internal
DNS, which it can get the same way every other zone does. (I left
`address` in place for consistency with other zones, but I was tempted
to remove it too: some discussion in #6651.)
* In sled-agent, we no longer try to resolve internal DNS via internal
DNS to populate `/etc/resolve.conf` for zones that expect to access
internal DNS; instead, we populate it with the fixed list of reserved
`INTERNAL_DNS_REDUNDANCY` IP addresses. This was immediately important
in getting tests passing, but is also probably the right thing to do; I
added a comment summarizing the points @davepacheco made in a call
earlier today.
* Renamed `MAX_INTERNAL_DNS_REDUNDANCY` to
`RESERVED_INTERNAL_DNS_REDUNDANCY`, and changed almost all of its users
to use `INTERNAL_DNS_REDUNDANCY` instead. This split was confusing, and
we had helper methods to construct DNS clients that disagreed on which
constant to use. `RESERVED_INTERNAL_DNS_REDUNDANCY` should be more clear
that this is reserved for future work/growth, and
`INTERNAL_DNS_REDUNDANCY` should be the value used in practice.
* Fixed a bug in the reconfigurator planner where the external
networking and DNS subnet allocators were considering all resources used
by the parent blueprint as in use, even if the planner had decided to
expunge the zones owning those resources. The fix for this is kinda
gross; I'm open to better ideas but also I'd like to take a crack at a
larger refactoring once some of the other big planner PRs have landed.

I had tested an R10 -> this branch upgrade on london, but only about
halfway through the above work. I want to retest both an R10 -> branch
upgrade and a fresh RSS install to ensure there are no lingering
expectations on the old internal NTP config path, but since I'd made it
through a successful upgrade, I think this can go ahead and be opened
for review.