Add a boot device field to Instance, forward along as appropriate (#6585)
the Omicron side of adding explicit boot order selection to instances
(counterpart to
[propolis#756](https://github.com/oxidecomputer/propolis/pull/756)).

first, this extends `params::InstanceCreate` to take a new `boot_disk:
Option<params::InstanceDiskAttachment>`.

additionally, this adds a `PUT /v1/instances/{instance}` to update
instances. the only property that can be updated at the moment is
`boot_disk`, pick a new boot disk or unset it entirely. this also
partially subsumes https://github.com/oxidecomputer/omicron/pull/6321.

finally, this updates Omicron to reference a recent enough Propolis that
#756 is included.

a surprising discovery along the way: you can specify a disk to be
attached multiple times in `disks` today, when creating an instance, and
we're fine with it! this carries through with the new `boot_disk` field:
if you specify the disk as `boot_disk` and in `disks`, it is technically
listing the disk for attachment twice but this will succeed.