[nexus] Make the default auto-restart policy real (#6697)
Instances created without an auto-restart policy are *supposed* to
default to the `BestEffort` policy, permitting them to reincarnate.
Unfortunately, however, **I am incredibly stupid** and, uh...I forgot to
make the database query for listing instances in need of reincarnation
handle instances whose `auto_restart_policy` is `NULL`. The in-memory
version of the check that an instance's policy permits it to reincarnate
honors the default, but because the database query does not include such
instances, they will never be reincarnated.

Making matters worse, none of our tests caught this, because we never
actually exercised the default policy --- when I wrote this code, we
hadn't settled on what the default would *be* yet, so all my tests for
it set explicit policies. Whoopsie!

This commit fixes that, and also actually adds a test so this doesn't
happen in the future. Good job, Eliza!