Start integrating clickhouse clusters to blueprints (#6627)
This PR integrates integrates clickhouse keeper and server node
allocation via `ClickhouseAllocator` into the `BlueprintBuilder`. It
goes about testing that allocation works as expected via the planner.
There are many more specific tests for the `ClickhouseAllocator` and the
tests here are mainly checking that everything fits together and the
same results occur when run through the planner. There is an additional
test and code to allow us to disable the cluster policy and expunge 
all clickhouse keeper and server zones in one shot.

This code is safe to merge because it is currently inert. There is no
way to enable the `ClickhousePolicy` outside of tests yet. This will
come in one or two follow up PRs where we add an internal nexus endpoint
for enabling the policy and then an OMDB command to trigger the
endpoint. Further OMDB support will be added for monitoring. We expect
that for the foreseeable future we will always deploy with
`ClickhousePolicy::deploy_with_standalone = true`. This is stage 1 of
RFD 468 where we run replicated clickhouse and the existing single node
clickhouse together. 

Lastly, there will be other PRs to plug in the actual inventory
collection and execution phases for clickhouse cluster reconfiguration.
We shouldn't bother even implementing the OMDB policy enablement until
all that is complete as it just won't work.