try running tests in release rather than debug in CI
this is a mixed bag: no debug_assertions but, in theory, should complete
significantly faster overall.

locally the timing breakdown works out like this:

debug:
  build:
    5m37s
  test:
    14m10s

release:
  build:
    11m20s
  test:
    5m43s

so in total, the extra time doing a release build (~6 minutes) is offset
by the time saved in tests (~8.5 minutes).

since this is such a small time save, perhaps it won't matter so much in
CI, but lets see!