Some extra route RPW version handling.
The place we were missing router version bumps was CRUD operations on
Internet Gateways themselves. The rationale here is that these are
resolved by name, so we need to indicate to the RPW that there may be a
new resource that can actually be resolved.

Returns one `continue` back into active service, since it appears to be
behaving well.