Correctly handle multiple IGW mappings in OPTE
It's currently possible for several IGWs in a VPC to be linked to the same IP
pool -- we now report all such mappings to OPTE and sled-agent.
OPTE now allows a given address to be selected for any IGW which would
admit it.

Previously, this led to a case where an OPTE port would lack usable NAT
entries for the IGW it was actually configured to use (due to the
presence of a second matching IGW).