Don't provision clickhouse servers & keepers in RSS (#6767)
Clickhouse clusters are only provisioned via reconfigurator as of
https://github.com/oxidecomputer/omicron/pull/6627.

This drastically simplifies things since we don't have to put more
policy knobs in RSS and wicket. It also matches the fact that these
zones always need nexus running in order to get their configurations and
run the appropriate processes anyway.

This commit removes the vestigal zone creation that was never actually
used.