[2/n] [reconfigurator] handle empty zones/disks use cases within ExampleSystem (#6768)
Currently, we build example systems and then mutate the inputs in various ways,
usually to empty out zones and disks.

With this PR, more of these cases are handled within the example system
via a builder pattern. This simplifies upcoming work to improve
`ExampleSystem` instances for more robust tests.

This also makes a change to `ExampleSystem` to not carry around the
blueprint -- logically, the blueprint evolves separately from the
underlying system. This change also simplifies some upcoming work around
making the collection and input behave correctly.