fix bug due to idempotency
The sled-agent instance-ensure-unregistered API is idempotent, which is
right and proper...but it introduces an issue in the case of a
forgotten VMM after a sled-agent restart. The
instance-ensure-unregistered call will return `None` because sled-agent
doesn't know about that instance, but if this is the first time we have
discovered that sled-agent doesn't know about the instance, we will need
to move it to `Failed`. This commit fixes that.