[3/n] [reconfigurator-cli] drop dependency on reconfigurator-execution (#6783)
The only functions from reconfigurator-execution currently pulled in are these
helper methods. Move them into nexus-types so that reconfigurator-cli no longer
has to pull in the execution crate and all of its dependencies (particularly
nexus-db-model and nexus-db-queries).

There is one downside, which is that `Overridables`'s `#[cfg(test)]` methods
are no longer marked that way. However, it does make the overrides available to
places such as the reconfigurator-cli. (It would be interesting to have a
test-related overlay feature in omicron, i.e. a `testing` feature in many of
our crates, but we don't have that yet.)

I had to move some of the silo fixed data into nexus-types as well -- as part
of that I took the opportunity to switch to const methods for a couple of
UUIDs.