Fix IP address create on Internet Gateway.
During testing we observed an issue with IGW pool linking on the CLI
where the error 'gateway was not initialized' cropped up. As it turns
out, this was down to a duplicated (and unnecessary) gateway parameter
in both the query string and the request body.

We had missed this on IP address create for individual IP address
creation/linking. This PR quickly removes this parameter from
`InternetGatewayIpAddressCreate`, and fixes some copy-paste description
mismatches for the same APIs.