[internal-dns] break types-only dependencies to dns-service-client (#6807)
Currently, a number of crates across Omicron use `dns-service-client` just for
the types, leading tools like `ls-apis` to have to make exceptions to handle
them.

To resolve this, follow the pattern we've used in other places, particularly
while breaking dependency loops between API and client crates:

* Introduce a `types` crate, in this case `internal-dns-types`, and move the
  shared types (currently in `dns-server-api`) to the crate. This crate is a
  good place to put config code generally.
* Use the types from `internal-dns-types` everywhere via `replace`
  directives.

As a result of this, the only code that was left in `internal-dns` was the
resolver -- so rename the crate to `internal-dns-resolver` and move that to
`internal-dns/resolver`. Also move the `internal-dns-cli` crate to
`internal-dns/cli`, since that is the general pattern we follow in
Omicron.

As a result of this change, code that just needs the types no longer needs to
depend on `dns-service-client`, allowing us to remove one of the two
`dns-service-client`-related rules. (I believe the other rule can be removed
once all our dependencies have been bumped.)

There are no changes to the output of `cargo xtask ls-apis apis`.